Fix marker rendering Remove old behavior chart
